Move auxiliary file client to a shared location (#583)
AuxiliaryFileClient is currently only used by search service. Now it will be used by Auxiliary2AzureSearch as well.
Progress on https://github.com/NuGet/NuGetGallery/issues/6447
