+++ src/NuGet.Jobs.Auxiliary2AzureSearch/Job.cs
@@ -6,8 +6,10 @@ using System.Threading.Tasks;
 using Autofac;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
+using Microsoft.Extensions.Options;
 using NuGet.Services.AzureSearch;
 using NuGet.Services.AzureSearch.Auxiliary2AzureSearch;
+using NuGet.Services.AzureSearch.AuxiliaryFiles;
 
 namespace NuGet.Jobs
 {
@@ -35,6 +37,8 @@ namespace NuGet.Jobs
             services.AddAzureSearch();
 
             services.Configure<Auxiliary2AzureSearchConfiguration>(configurationRoot.GetSection(ConfigurationSectionName));
+            services.AddTransient<IOptionsSnapshot<IAuxiliaryDataStorageConfiguration>>(
+                p => p.GetRequiredService<IOptionsSnapshot<Auxiliary2AzureSearchConfiguration>>());
             services.Configure<AzureSearchJobConfiguration>(configurationRoot.GetSection(ConfigurationSectionName));
             services.Configure<AzureSearchConfiguration>(configurationRoot.GetSection(ConfigurationSectionName));
         }
