+++ tests/NuGet.Services.AzureSearch.Tests/SearchService/AuxiliaryDataFacts.cs
@@ -4,6 +4,7 @@
 using System;
 using System.Collections.Generic;
 using NuGet.Indexing;
+using NuGet.Services.AzureSearch.AuxiliaryFiles;
 using Xunit;
 
 namespace NuGet.Services.AzureSearch.SearchService
